only build pushes to master
If a developer opens a pull request on a feature branch, it means that CI takes twice as much time because it builds both the feature branch and the pull request.
Most of the time we are only interested in testing the result after the merge rather than the state of the current feature branch.